BookManagementSystem/src/test/java/main/controller/UserControllerTest.java
package main.controller;

import com.fasterxml.jackson.databind.ObjectMapper;
import main.dto.UserDto;
import main.service.UserService;
import main.utils.UsersMocks;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.client.match.MockRestRequestMatchers.jsonPath;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(controllers = UserController.class)
public class UserControllerTest {
    @MockBean
    private UserService userService;

    @Autowired
    private MockMvc mockMvc;

    @Autowired
    ObjectMapper objectMapper;


    @Test
    public void TestAddUser() throws Exception{
        // Arrange
        UserDto userDto = UsersMocks.mockUserDto();
        when(userService.addUser(any())).thenReturn(userDto);

        String userDtoBody = objectMapper.writeValueAsString(userDto);

        MvcResult result = mockMvc.perform(post("/users")
                        .content(userDtoBody)
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andReturn();
        assertEquals(result.getResponse().getContentAsString(), userDtoBody);
    }
}
